requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
Django==4.2.3
django-environ==0.10.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.98.0
h11==0.14.0
idna==3.4
pydantic==1.10.11
PyJWT==2.7.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0
